requirements.txt
astropy>=5.3
requests>=2.31
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
beautifulsoup4>=4.12
lxml>=4.9
tqdm>=4.65
aiohttp>=3.9
tenacity>=8.2
python-dateutil>=2.8
